attempting to fix session bug
my final attempt to fix the session bug. caused by my choice to use the ---api tag when generating the rails backend. after a full day of research and debugging, I have to redo the backend part of this project. Will be deleting the backend file and creating a new backend. I was able to successfully reimplement session cookies into the API, but im unable to set custom session cookies such as session[:user_id]. Rails simply wont let me due to much of the middleware being left out during generation. The next push will be a big one, but only for the backend.