Bug fixes in forms & controllers
Fixed error views in the login and signup forms while also removing some unnecessary methods from the controllers. Project could be submitted like this, as this is stable with all the remaining known bugs fixed. One last feature I want to implement after this is the search and add friends.