Added friends list & cleaned controllers
Added a friends list to the current users profile page, and also filled out the UserProfile page for the friends. I also got ride of unnecessary instance variables and made them local variables. Only need to fix some bugs in the routing and the store losing the friendsReducer when ObservationReducer is called.