Refactored Search Bar
Took out unnecessary code out from the search bar, such as  setting the results to an empty array when no value (happens already with setResults) and handleResultSelect. This method doesnt matter because the user is redirected anyway after selecting the result and doesnt need the user they clicked on in the search bar after clicking the link.